Change display types for divs
Changed the display of the container to flex and
allow the container to wrap; also add width and
height to container instead of body. Removed
margin in container to uncenter it as well. Added
width to row to ensure the rows take up full
length and flex: 1 to column to allow the columns
to grow and fill space